+++ src/main/java/io/pinecone/clients/Index.java
@@ -9,7 +9,9 @@ import io.pinecone.proto.*;
 import io.pinecone.proto.DeleteRequest;
 import io.pinecone.proto.DescribeIndexStatsRequest;
 import io.pinecone.proto.FetchResponse;
+import io.pinecone.proto.ListNamespacesResponse;
 import io.pinecone.proto.ListResponse;
+import io.pinecone.proto.NamespaceDescription;
 import io.pinecone.proto.QueryRequest;
 import io.pinecone.proto.UpdateRequest;
 import io.pinecone.proto.UpsertRequest;
@@ -50,7 +52,9 @@ public class Index implements IndexInterface<UpsertResponse,
         UpdateResponse,
         DeleteResponse,
         DescribeIndexStatsResponse,
-        ListResponse> {
+        ListResponse,
+        ListNamespacesResponse,
+        NamespaceDescription> {
 
     private final PineconeConnection connection;
     private final String indexName;
@@ -976,6 +980,87 @@ public class Index implements IndexInterface<UpsertResponse,
         ListRequest listRequest = ListRequest.newBuilder().setNamespace(namespace).setPrefix(prefix).
                 setPaginationToken(paginationToken).setLimit(limit).build();
         return blockingStub.list(listRequest);
+    }
+
+    /**
+     * <pre>
+     * Overload to get a list of all namespaces withing an index without limit and pagination token. When limit is not
+     * set, it'll default to 100.
+     * @return {@link ListNamespacesResponse} The response for the list namespace operation.
+     * </pre>
+     */
+    @Override
+    public ListNamespacesResponse listNamespaces() {
+        ListNamespacesRequest listNamespacesRequest = ListNamespacesRequest
+                .newBuilder()
+                .build();
+        return blockingStub.listNamespaces(listNamespacesRequest);
+    }
+
+    /**
+     * <pre>
+     * Overload to get a list of all namespaces withing an index with default limit set to 100.
+     * @param paginationToken The token to paginate through the list of vector IDs.
+     * @return {@link ListNamespacesResponse} The response for the list namespace operation.
+     * </pre>
+     */
+    @Override
+    public ListNamespacesResponse listNamespaces(String paginationToken) {
+        ListNamespacesRequest listNamespacesRequest = ListNamespacesRequest
+                .newBuilder()
+                .setPaginationToken(paginationToken)
+                .build();
+        return blockingStub.listNamespaces(listNamespacesRequest);
+    }
+
+    /**
+     * <pre>
+     * Get list of all namespaces within an index.
+     * @param paginationToken The token to paginate through the list of vector IDs. If pagination token is set to null,
+     *                        it'll be ignored.
+     * @param limit           The maximum number of vector IDs you want to retrieve.
+     * @return {@link ListNamespacesResponse} The response for the list namespace operation.
+     * </pre>
+     */
+    @Override
+    public ListNamespacesResponse listNamespaces(String paginationToken, int limit) {
+        ListNamespacesRequest.Builder listNamespacesRequest = ListNamespacesRequest
+                .newBuilder()
+                .setLimit(limit);
+        if(paginationToken != null) {
+            listNamespacesRequest.setPaginationToken(paginationToken);
+        }
+        return blockingStub.listNamespaces(listNamespacesRequest.build());
+    }
+
+    /**
+     * <pre>
+     * Describe a namespace within an index, showing the vector count within the namespace.
+     * @param namespace The namespace to describe.
+     * @return {@link NamespaceDescription} The response for the describe namespace operation.
+     * </pre>
+     */
+    public NamespaceDescription describeNamespace(String namespace) {
+        DescribeNamespaceRequest describeNamespaceRequest = DescribeNamespaceRequest
+                .newBuilder()
+                .setNamespace(namespace)
+                .build();
+        return blockingStub.describeNamespace(describeNamespaceRequest);
+    }
+
+    /**
+     * <pre>
+     * Delete a namespace from an index.
+     * @param namespace The namespace to describe.
+     * @return {@link DeleteResponse} The response for the delete namespace operation.
+     * </pre>
+     */
+    public DeleteResponse deleteNamespace(String namespace) {
+        DeleteNamespaceRequest deleteNamespaceRequest = DeleteNamespaceRequest
+                .newBuilder()
+                .setNamespace(namespace)
+                .build();
+        return blockingStub.deleteNamespace(deleteNamespaceRequest);
     }
 
     /**