+++ src/main/java/io/pinecone/commons/Constants.java
@@ -1,5 +1,5 @@
 package io.pinecone.commons;
 
 public class Constants {
-    public static final String pineconeClientVersion = "v4.0.1";
+    public static final String pineconeClientVersion = "v5.0.0";
 }